Moved GameFX out of "drawing" package
